requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
annoy==1.17.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.103.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.75
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nemoguardrails==0.6.1
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.5.0
packaging==23.2
Pillow==10.1.0
pydantic==1.10.9
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
simpleeval==0.9.13
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==2.1.0
uvicorn==0.23.2
yarl==1.9.4
